actions: Split out zerver.actions.user_settings.
(cherry picked from commit ec6355389a3e586c5180993a4ce4e08c9d583b38)
