+++ zerver/forms.py
@@ -19,8 +19,9 @@ from markupsafe import Markup as mark_safe
 from two_factor.forms import AuthenticationTokenForm as TwoFactorAuthenticationTokenForm
 from two_factor.utils import totp_digits
 
+from zerver.actions.user_settings import do_change_password
 from zerver.decorator import rate_limit_request_by_ip
-from zerver.lib.actions import do_change_password, email_not_system_bot
+from zerver.lib.actions import email_not_system_bot
 from zerver.lib.email_validation import email_allowed_for_realm
 from zerver.lib.exceptions import JsonableError, RateLimited
 from zerver.lib.name_restrictions import is_disposable_domain, is_reserved_subdomain
