+++ zerver/management/commands/bulk_change_user_name.py
@@ -3,7 +3,7 @@ from typing import Any
 
 from django.core.management.base import CommandError
 
-from zerver.lib.actions import do_change_full_name
+from zerver.actions.user_settings import do_change_full_name
 from zerver.lib.management import ZulipBaseCommand
