+++ zerver/tests/test_audit_log.py
@@ -8,29 +8,31 @@ from django.utils.timezone import now as timezone_now
 from analytics.models import StreamCount
 from zerver.actions.realm_icon import do_change_icon_source
 from zerver.actions.realm_playgrounds import do_add_realm_playground, do_remove_realm_playground
+from zerver.actions.user_settings import (
+    do_change_avatar_fields,
+    do_change_password,
+    do_change_tos_version,
+    do_change_user_delivery_email,
+    do_change_user_setting,
+    do_regenerate_api_key,
+)
 from zerver.actions.users import do_change_user_role, do_deactivate_user
 from zerver.lib.actions import (
     bulk_add_subscriptions,
     bulk_remove_subscriptions,
     do_activate_mirror_dummy_user,
     do_add_realm_domain,
-    do_change_avatar_fields,
     do_change_bot_owner,
     do_change_default_all_public_streams,
     do_change_default_events_register_stream,
     do_change_default_sending_stream,
-    do_change_password,
     do_change_realm_domain,
     do_change_subscription_property,
-    do_change_tos_version,
-    do_change_user_delivery_email,
-    do_change_user_setting,
     do_create_user,
     do_deactivate_realm,
     do_deactivate_stream,
     do_reactivate_realm,
     do_reactivate_user,
-    do_regenerate_api_key,
     do_remove_realm_domain,
     do_rename_stream,
     do_set_realm_authentication_methods,
