+++ zerver/tests/test_push_notifications.py
@@ -23,10 +23,10 @@ from requests.models import PreparedRequest
 
 from analytics.lib.counts import CountStat, LoggingCountStat
 from analytics.models import InstallationCount, RealmCount
+from zerver.actions.user_settings import do_regenerate_api_key
 from zerver.lib.actions import (
     do_delete_messages,
     do_mark_stream_messages_as_read,
-    do_regenerate_api_key,
     do_update_message_flags,
 )
 from zerver.lib.avatar import absolute_avatar_url
