+++ zerver/views/registration.py
@@ -26,6 +26,11 @@ from confirmation.models import (
     validate_key,
 )
 from zerver.actions.default_streams import lookup_default_stream_groups
+from zerver.actions.user_settings import (
+    do_change_full_name,
+    do_change_password,
+    do_change_user_setting,
+)
 from zerver.context_processors import get_realm_from_request, login_context
 from zerver.decorator import do_login, rate_limit_request_by_ip, require_post
 from zerver.forms import (
@@ -35,14 +40,7 @@ from zerver.forms import (
     RealmRedirectForm,
     RegistrationForm,
 )
-from zerver.lib.actions import (
-    do_activate_mirror_dummy_user,
-    do_change_full_name,
-    do_change_password,
-    do_change_user_setting,
-    do_create_realm,
-    do_create_user,
-)
+from zerver.lib.actions import do_activate_mirror_dummy_user, do_create_realm, do_create_user
 from zerver.lib.email_validation import email_allowed_for_realm, validate_email_not_already_in_realm
 from zerver.lib.exceptions import RateLimited
 from zerver.lib.pysa import mark_sanitized
